+++ 2018-08_AkkaIntro/code/Supervision/ProcessANumber.cs
@@ -1,0 +1,12 @@
+﻿namespace Supervision
+{
+    public class ProcessANumber
+    {
+        public int Number { get; }
+
+        public ProcessANumber(int number)
+        {
+            Number = number;
+        }
+    }
+}
