+++ 2018-08_AkkaIntro/code/Supervision/Program.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Linq;
 using Akka.Actor;
-using Akka.Routing;
 using DotNetty.Handlers.Timeout;
 using Shared.Actors;
-using Shared.Messages;
 
 namespace Supervision
 {
@@ -24,74 +22,6 @@ namespace Supervision
             }
 
             Console.ReadLine();
-        }
-    }
-
-    public class ParentActor : ReceiveActor
-    {
-        private readonly IActorRef _volatileChildren;
-
-        public ParentActor()
-        {
-            var routerSupervisionStrategy = new OneForOneStrategy(
-                localOnlyDecider: ex => Directive.Resume);
-
-            var props = Props.Create<VolatileChildActor>()
-                .WithRouter(new RoundRobinPool(1)
-                    .WithSupervisorStrategy(routerSupervisionStrategy)
-                );
-
-            _volatileChildren = Context.ActorOf(props, "children");
-
-            Receive<ProcessANumber>(msg =>
-            {
-                _volatileChildren.Tell(msg);
-            });
-        }
-    }
-
-    public class VolatileChildActor : ReceiveActor, IWithUnboundedStash
-    {
-        private readonly ActorSelection _consoleWriter;
-        private int processedNumbers = 0;
-        public VolatileChildActor()
-        {
-            _consoleWriter = Context.ActorSelection("/user/consoleWriter");
-
-
-            Receive<ProcessANumber>(msg =>
-            {
-                if (processedNumbers == 6)
-                {
-                    processedNumbers = 0;
-                    throw new Exception($"I already processed 6 numbers; too tired to process number {msg.Number}");
-                }
-
-                _consoleWriter.Tell(new WriteSomethingMessage($"Processing number: {msg.Number}"), Self);
-                processedNumbers++;
-            });
-        }
-
-        //protected override void PreRestart(Exception reason, object message)
-        //{
-        //    Stash.Stash();
-        //}
-
-        //protected override void PostRestart(Exception reason)
-        //{
-        //    Stash.UnstashAll();
-        //}
-
-        public IStash Stash { get; set; }
-    }
-
-    public class ProcessANumber
-    {
-        public int Number { get; }
-
-        public ProcessANumber(int number)
-        {
-            Number = number;
         }
     }
 }
